+++ src/util/countAVE.ts
